+++ lib/sbmt/outbox/instrumentation/open_telemetry_loader.rb
@@ -5,6 +5,7 @@ require "opentelemetry-common"
 require "opentelemetry-instrumentation-base"
 
 require_relative "../middleware/open_telemetry/tracing_create_item_middleware"
+require_relative "../middleware/open_telemetry/tracing_create_batch_middleware"
 require_relative "../middleware/open_telemetry/tracing_item_process_middleware"
 
 module Sbmt
@@ -15,6 +16,7 @@ module Sbmt
           require_dependencies
 
           ::Sbmt::Outbox.config.create_item_middlewares.push("Sbmt::Outbox::Middleware::OpenTelemetry::TracingCreateItemMiddleware")
+          ::Sbmt::Outbox.config.create_batch_middlewares.push("Sbmt::Outbox::Middleware::OpenTelemetry::TracingCreateBatchMiddleware")
           ::Sbmt::Outbox.config.item_process_middlewares.push("Sbmt::Outbox::Middleware::OpenTelemetry::TracingItemProcessMiddleware")
         end
 
@@ -26,6 +28,7 @@ module Sbmt
 
         def require_dependencies
           require_relative "../middleware/open_telemetry/tracing_create_item_middleware"
+          require_relative "../middleware/open_telemetry/tracing_create_batch_middleware"
           require_relative "../middleware/open_telemetry/tracing_item_process_middleware"
         end
       end